feat: Combat Screen Opening/Closing
Added code to the combat starter that opens the combat screen when starting combat and disables player movements whilst in the screen.
Currently leaving combat requires the user to press Q.
This will be replaced in later updates with the user having to complete the fight to exit the combat menu.