feat: edited enemy generator
Made it so that enemy generator spawns the enemy at its position.#
Removed and renamed prefabs and duplicate Enemy objects.
Changed enemy generator so that the type pick method uses switch case to be faster.
Changed the type of enemy to be a enum rather than a int;
Tweaked the Stats for the enemy types.