feat: made the menu for the combat screen work
Still needs the player skills to be added and the enemy turn stuff.
Removed the combat screen that was in the test environment as having two combat screen can interfere with each other.
Changed the colour of the buttons on mouseover.
Changed the code to add the onClick functions to the buttons after creating them in CombatScreen.cs.