feat: added quiz door and quiz trap door prefabs and started work on their scripts
Currently the doors use the quiz spawner prefab object to create the quiz and the quiz needs to run the new quizcleared() method after the quiz is over to remove the door and itself from the scene so the player can continue.
the trap door uses the code that was in the quiz spawner but will require some tweaks to fine tune it's functionality.